Minor test clean-up
