+++ requirements-dev.txt
@@ -1,4 +1,3 @@
 flake8
-gitpython >= 2.1.10, < 3
 mypy >= 0.600
 -r requirements.txt
